requirements.txt
Pillow
pandas
pandas
streamlit